Changes based on PR #2 comments
