+++ README.md
@@ -12,9 +12,9 @@ Prior to the installation, download the lastest version of TeraStitcher with sup
 Create a new conda environment and execute the following commands:
 
 ```
-$ conda create -n stitching python=3.8
+$ conda create -n stitching -y python=3.8
 $ conda activate stitching
-$ conda install ipykernel
+$ conda install -y ipykernel
 ```
 
 Then, open a terminal and clone this repository in the virtual machine and place the dowloaded `.tar.gz` file in the following path: `terastitcher-module/environment/GCloud/`. Now, execute the installer (this will take a couple of minutes):
@@ -51,7 +51,7 @@ $ python terastitcher.py --help
 
 This will be helpful if you want to overwrite them. 
 
-**Note**: It is important to mention that this module uses different tools that output processed images. For erasing streaks in EM images, we are using pystripe which output the same dataset folder structure with the processed images. Terastitcher does the same but with a different folder structure and at the end we convert the stitched images to OME-Zarr. Therefore, **you need 3X disk space based on your original disk size**. By default, we have a flag `clean_output` which is set to True. When this flag is set to True, it deletes the intermediate images generated by these module leaving only the OME-Zarr output.
+**Note**: It is important to mention that this module uses different tools that output processed images. For erasing streaks in EM images, we are using pystripe which outputs the same dataset folder structure with the processed images. Terastitcher does the same but with a different folder structure and finally we convert the stitched images to OME-Zarr. Therefore, **you need 3X disk space based on your original disk size**. By default, we have a flag `clean_output` which is set to True. When this flag is set to True, it deletes the intermediate images generated by this module leaving only the OME-Zarr output.
 
 If you run this pipeline in GCP, this module will automatically load the bucket(s) using GCSFUSE. As today, there are some performance limitations when using GCSFUSE with MPI if your dataset contains many small images. The performance will improve if you have larger files containing these small images (e.g. 3D stacks of images).
